Add removeIncompleteResponse
