+++ src/lib/Setting/Pages/AdvancedSettings.svelte
@@ -77,6 +77,9 @@
     <Check bind:check={$DataBase.autoContinueChat} name={language.autoContinueChat}> <Help key="autoContinueChat"/></Check>
 </div>
 <div class="flex items-center mt-4">
+    <Check bind:check={$DataBase.removeIncompleteResponse} name={language.removeIncompleteResponse}></Check>
+</div>
+<div class="flex items-center mt-4">
     <Check bind:check={$DataBase.newOAIHandle} name={language.newOAIHandle}/>
 </div>
 <div class="flex items-center mt-4">
